format the test
